Fix signed-shift UB on 16-bit-int targets
On platforms with 16-bit int (e.g. AVR, MSP430), uint8_t operands promote
to signed int, so shifting a byte >= 128 left by 8 overflows INT_MAX.
This affects mlk_rej_uniform_c and mlk_poly_frombytes_c; cast the shifted
operand to uint16_t to avoid the overflow.

The former was discovered by Claude Opus 4.8 during a security audit; the
latter was found after adding UBSan on AVR (see next commit).

Resolves #1726

+++ mlkem/src/compress.c
@@ -641,7 +641,12 @@ __contract__(
     const uint8_t t0 = a[3 * i + 0];
     const uint8_t t1 = a[3 * i + 1];
     const uint8_t t2 = a[3 * i + 2];
-    r->coeffs[2 * i + 0] = (int16_t)(t0 | ((t1 << 8) & 0xFFF));
+    /* Safety:
+     * - The explicit cast to uint16_t ensures that << 8 does
+     *   not signed-overflow even on a 16-bit system.
+     * - The cast to int16_t is safe due to the explicit 0xFFF truncation.
+     */
+    r->coeffs[2 * i + 0] = (int16_t)(t0 | (((uint16_t)t1 << 8) & 0xFFF));
     r->coeffs[2 * i + 1] = (int16_t)((t1 >> 4) | (t2 << 4));
   }
